Master moved from 97c7d0e -- where this branch last merged it -- to f7f9665, and
the three changes that landed in between touch exactly the files upstream also
rewrote. Git resolved the text without a single conflict, which is the outcome
that needs the most checking rather than the least: an automatic merge only says
the hunks did not overlap, not that the fork's invariants survived. Each one was
therefore verified against the merged tree rather than assumed.

  - cbb3710 moved CNetworkAddress off boost::asio::ip::address. Post-merge scan:
    boost/asio is reached by 4 of 274 translation units under src/ (3 distinct
    sources -- LibSocketAsio.cpp, NetworkAddress.cpp, webapi/HttpServer.cpp;
    LibSocket.cpp counts because it includes LibSocketAsio.cpp), unchanged from
    master's 4 of 255. ngtcp2 and gnutls each reach 1 TU, QuicLibraryAdapter.cpp.
    No public header regained an asio include.
  - 990c547 replaced FindClientByIP with FindClientByUDPEndpoint at the two NAT
    rendezvous sites. src/ClientUDPSocket.cpp is byte-identical to master, both
    call sites still use the endpoint lookup, and the file contains no
    FindClientByIP call at all.
  - 8ac9002 / 27dff3d added the QUIC transport and its hello-tag authentication.
    The CT_MOD_QUIC_IDENT parse, RegisterQuicExpectation and the two capability
    bits are intact in BaseClient.cpp; LibSocket.h keeps its IStreamTransport
    accessors, to which upstream only appended EnableTcpNoDelay.

Upstream's own restructuring is adopted wholesale where it does not collide with
fork semantics: CBrowseManager replaces the browse state that lived on
CUpDownClient and CClientList (the fork's m_pendingBrowses list and its sweep are
gone as dead code, upstream owns the lifecycle now), EContactResult names what
TryToConnect's bool used to conflate, and the search-restore moves after
LoadMetFiles.

The one place upstream's structure needed the fork's semantics ported into it is
CheckContactPreconditions(). Upstream extracted the connect preconditions but
kept the narrow `if (!uClientIP) return Contacting;` guard, which lets a native
IPv6 peer past both the ipfilter and the ban check -- the exact defect the
address widening fixed. The extraction is kept; the widened lookup runs inside
it, trying m_userAddress, then the hybrid ID via FromIPv4HostOrderOrAbsent()
when !HasLowID(), then m_connectAddress, and returning Contacting only when all
three are absent.

src/extern/libutp is untouched: 20 of 20 files still byte-identical to
transmission/libutp at 490874c44a2ecf914404b0a20e043c9755fff47b. No -Werror flag
in src/CMakeLists.txt was weakened, removed or given an exclusion.

+++ src/SearchFile.cpp
@@ -223,7 +223,7 @@ bool CSearchFile::WriteToFile(CFileDataIO *file) const
 		file->WriteUInt32(client.m_ip);
 		file->WriteUInt16(client.m_port);
 		file->WriteUInt32(client.m_serverIP);
-		file->WriteUInt16((uint16)client.m_serverPort);
+		file->WriteUInt16(client.m_serverPort);
 	}
 
 	file->WriteUInt16((uint16)m_children.size());
